Chore:updated quiz progress
