+++ API/src/controllers/course.js
@@ -12,8 +12,6 @@ const {
   toggleAvailablity,
   toggleEditing,
   evaluateUserAnswers,
-  getAUserCourse,
-  updateScore,
 } = require("../services/course");
 
 const createCourse = async (req, res) => {
@@ -83,24 +81,6 @@ const getAllCourses = async (req, res) => {
     return res.status(200).send({
       success: true,
       data: filtered,
-    });
-  } catch (error) {
-    return res.status(400).send({
-      success: false,
-      message: "Request failed",
-    });
-  }
-};
-
-const getUserCourse = async (req, res) => {
-  try {
-    const courseId = req.params.courseId;
-
-    const course = await getAUserCourse(req.user.id, courseId);
-
-    return res.status(200).json({
-      success: true,
-      data: course,
     });
   } catch (error) {
     return res.status(400).send({
@@ -254,30 +234,10 @@ const toggleCourseEditing = async (req, res) => {
 
 const evaluateQuizScore = async (req, res) => {
   const quizPayload = req.body;
-  const userId = req.params.userId;
+
   const courseId = req.params.courseId;
   try {
-    const score = await evaluateUserAnswers(userId, courseId, quizPayload);
-
-    return res.status(200).json({
-      success: true,
-      score: score,
-    });
-  } catch (error) {
-    console.log(error);
-    return res.status(400).send({
-      success: false,
-      message: "Request failed",
-    });
-  }
-};
-
-const updateQuizScore = async (req, res) => {
-  const quizPayload = req.body;
-  const userId = req.params.userId;
-  const courseId = req.params.courseId;
-  try {
-    const score = await updateScore(userId, courseId, quizPayload);
+    const score = await evaluateUserAnswers(req.user.id, courseId, quizPayload);
 
     return res.status(200).json({
       success: true,
@@ -306,6 +266,4 @@ module.exports = {
   toggleCourseAvailablity,
   toggleCourseEditing,
   evaluateQuizScore,
-  updateQuizScore,
-  getUserCourse,
 };
