+++ API/src/routes/user.js
@@ -1,9 +1,9 @@
 const router = require("express").Router();
 
-const { getUserOverAllScore } = require("../controllers/user");
+const { getUserOverAllScore, getUserCourse } = require("../controllers/user");
 const { basicAuth } = require("../middlewares/auth");
 
 router.use(basicAuth());
 router.get("/overall-score/:courseId", getUserOverAllScore);
-
+router.get("/course/:courseId", getUserCourse);
 module.exports = router;
